Move view-classes to corresponding BEM blocks folders
